requirements.txt
albumentations==0.4.1
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.0
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
imageio==2.6.1
imgaug==0.2.6
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==7.2.0
networkx==2.4
numpy==1.17.3
opencv-python-headless==4.1.1.26
packaging==19.2
Pillow==6.2.0
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.2.1
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
requests==2.22.0
scikit-image==0.16.1
scipy==1.3.1
six==1.12.0
supervisor==4.1.0
torch==1.3.0.post2
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
zipp==0.6.0
